updated part 1
