+++ db/seed.js
@@ -43,65 +43,3 @@ let breakfasts = [Breakfast1, Breakfast2, Breakfast3]
 //   })
 // .then(disconnect)
 // } 
-
-
-// I will wrap them in a function create for usability
-const create = (name, address, yelpUrl, items) => {
-    const restaurantParams = {
-        name: name,
-        address: address,
-        yelpUrl: yelpUrl,
-        items:  items 
-        }
-    RestaurantModel.create(restaurantParams)
-    .then((restaurant) => {
-        console.log(restaurant)
-    })
-    .catch((error) => console.log(error))
-    .then(disconnect)
-     
-
-}   // end create func  
-
-// create("Burger King", {street: "ALkhaleej", zipcode: 555555}, "IDK", breakfasts)
-
-
-const find = (restaurantName) => {
-    RestaurantModel.findOne({name: restaurantName}) // a condition where the key name is equal to the parameter restaurantName
-    .then( (restaurant) => {
-        console.log(`You are looking for ${restaurant} restaurant`)
-    })
-    .catch((error) => console.log(error))
-    .then(disconnect)
-} // end func
-
-// find("Urth")
-
-
-const findAll = (zipcodeInput) => {
-    RestaurantModel.find({zipcode: zipcodeInput})
-    .then( (restaurants) => {
-        console.log(restaurants)
-    })
-    .catch( (error) => console.error)
-    .then(disconnect)
-} // end func
-
-// findAll(11111)
-
-const update = (rest_id, updates) => {
-        RestaurantModel.findByIdAndUpdate(rest_id, updates)
-        .then((student) => {
-            console.log(student)
-        })
-        .catch( (error) => console.error)
-        .then(disconnect)
-} // end func
-
-const destroy = (rest_id) => {
-    RestaurantModel.findByIdAndRemove(rest_id)
-  .then(findAll)
-  .catch( (error) => console.error)
-  .then(disconnect)
-
-} // end func
